Merged stacked graph support from iownbey's version and added assorted fixes to 'unstack' them when calling the highlightCallback.
Added an option to not hide the label on mouse out.

Alternate the list of auto-generated colors since adjacent colors
are difficult to distinguish. A better way would be to randomize them with a
constant seed, but alas, Javascript's Math.random() doesn't accept a seed.

Added a getColors() accessor method to help an external labelsDiv
obtain the list of [auto-generated] colors.