From antrob; don't recalculate layout when updateOptions() would not change it.
This suppresses recalculation of the pixel position of points when calling
updateOptions() with options which will not affect the layout. Insignificant
for small graphs, but this change provides a great performance boost for
changing options for graphs with multiple data sets.

Anthony saw something like a 30-40% speedup for dense charts.