Various fixes/workarounds for IE7/IE8 and some fixes/tweaks of range selector
For background on IE7/IE8 work, see the following:

http://groups.google.com/group/dygraphs-users/browse_thread/thread/8c305992bf185ea8

Also, I have a fix of a recent commit in dygraph-utils.js that checks for instance of Node, which doesn't exist in IE7/IE8.

In addition to some minor tweaks and code cleanup in dygraph-range-selector.js, there's a fix for issue #90.