Fix issue 131: setVisibility() broken with multiple axes
This reworks the logic for visibility and axes. Invisible series are kept in
the seriesToAxis map, but are ignored when calculating the layout and extreme
values.

I am slightly wary because I did not understand the comment about visibility in
the deleted portion of computeYAxes_. It seemed to be describing the exact same
problem as this bug, but did not really address it.

Squashed commit of the following:

commit 5916f4782d26e0933764607dd7cc93ebde981b2a
Author: Dan Vanderkam <dan@dygraphs.com>
Date:   Wed Dec 21 16:26:56 2011 -0500

    I do not believe so

commit ae08301858225d9c9b2eb185cb45be8a3cf1387e
Author: Dan Vanderkam <dan@dygraphs.com>
Date:   Wed Dec 21 16:25:52 2011 -0500

    issue is fixed, but are there side-effects?

commit 4cfbfae182d800bd15c738e879c707b283117983
Merge: c26c4fd 107c99e
Author: Dan Vanderkam <dan@dygraphs.com>
Date:   Wed Dec 21 15:58:24 2011 -0500

    Merge branch 'master' into issue131

commit c26c4fd945d1b515d2ebe7839fb5fc7a9cfa2cec
Author: Dan Vanderkam <dan@dygraphs.com>
Date:   Wed Dec 21 15:51:03 2011 -0500

    failing auto_test