Add support for closest-series highlighting
The new option highlightSeriesOpts applies options such as strokeWidth
to the timeseries that's currently closest to the mouse Y position,
and also passes the name of that timeseries to the highlightCallback.

Also add a "highlight" class label to the currently-highlighted
timeseries in the legend view to facilitate styling via CSS.

See the new tests in callback.js for examples.