Add support for closest-series highlighting
The new option highlightSeriesOpts applies options such as strokeWidth
to the timeseries that's currently closest to the mouse Y position,
and also passes the name of that timeseries to the highlightCallback.

There's also a new optional seriesName argument to setSelection().

See the new tests in callback.js for examples.

Also add a "highlight" class label to the currently-highlighted
timeseries in the legend view to facilitate styling via CSS.

The new options strokeBorderWidth and strokeBorderColor allow drawing a
border around graph lines to make them stand out better. This is
intended for highlightSeriesOpts, but also works when used for all
lines.

There's also support for background fading for series highlighting,
optionally animated.

New options:
  highlightSeriesOpts (default null)
  strokeBorderWidth (default 0 == off)
  strokeBorderColor (default white)
  highlightSeriesBackgroundFade (default 0 == off)
  highlightSeriesAnimated (default false)