Keep the X axis ticks when logscale is set globally.
Currently, setting {logscale: true} at the toplevel options with a
numeric X axis attempts to use logarithmic ticks for the X axis too,
with strange results since the X values aren't scaled logarithmically.
This resulted in tests/logscale.html skipping the "6" and "8" numbers on
the ticker, and the modified demo with a large X value shows the effect
more clearly.

This change wraps the numericTicks generator as numericLinearTicks,
ignoring 'logscale' in its axis options, and uses that for the X axis
numeric ticker. Arguably doing it the other way around would be clearer,
but that would be a much larger refactoring.

(Note that date-valued X axes as used in the stock demo weren't affected
by this issue.)