Fix highlighting for zoomed graphs
The code wasn't correctly applying the left boundary offset when converting a row index in the pruned data to/from a row index for the full data set.

Add a getLeftBoundary_ helper method and replace the existing loops that attempted to find the first defined record. Not sure why it saves multiple boundaryIds anyway, the code assumes it can just use the first one.

Also fix an apparently-invalid usage in getSelection().