Fix connectSeparatedPoints, add test
The original connectSeparatedPoints code renumbered points within series, with the effect that 'row' as used in setSelection didn't work as expected - it selects unrelated points.

This change keeps the point numbering intact, and instead moves the logic to skip missing points to the canvas renderer and setSelection().

A new auto_test confirms this behavior.