Fix NaN handling for stacked graphs.
Dygraph was handling NaNs inconsistently, drawing lines and shaded areas
for values encountered while stacking until it ran into a NaN, then
stopping at that point due to NaN contagion, with the effect that it
drew some lines and partial gaps.

Change this to create a clean gap in the data by propagatng the NaN contagion
in both directions.

Note that users can preprocess the data to replace NaNs with zeros if
they prefer that, but this should not happen by default since it's
irreversible and can lead to confusing graphs that don't match actual
data.