Add optional "locked" arg to setSelection
Locking a series selection temporarily disables the closest-series
logic of the highlighting code, switching it to traditional by-row
selection until unlocked by clearSelection().

This is useful for following values of a single series on mouseover,
see the example use in the "Highlight Closest Series" demo which
toggles this by clicking the graph.