Export .isSeriesLocked() API function
This is a followup to https://github.com/danvk/dygraphs/pull/159,
exposing the .lockedSet_ internal attribute through the API
for use in callbacks, and also with an eye for future extraction
of the mouseover logic into a plugin.

Includes a fix for the auto_test name, it was erroneously replacing
the existing testClosestPointCallbackCss1 test due to a name clash.