Y-axis labels: hidden overflow fix
I know there is a (y)axisLabelWidth option, but in our application we don't know what for data the user will load into the dygraph, so we can't really put a fixed value on this.

Therefor it would be better if dygraph would always fully show the values in the Y-axis, no matter what the axisLabelWidth might be. Otherwise it can create situations like this: http://i48.tinypic.com/wqvacx.jpg in this graph the amount of handles of a particular process is shown, but it should be 60000 handles instead of 600. Also in our application some values have a special value and are being replaced by text, as you can see in the graph at the bottom, where it should say 'Not initialized' instead of 'Not'.

In this patch I also added a backgroundcolor option, so when the labels would overlap with the axis/graph the label would have a background, just like the mouse over legend overlay.