LogScale and customBars with negative values
When LogScale and customBars are enabled, it was verifying if the value was negative by using a <= on an array and then inserting a null. This caused exceptions in rollingAverage(), the value can't be null when using customBars.

(This a new patch based on the current code, since pull request #236 was getting old).