Track stroke/fill in CanvasAssertions.
For lineTo operations, set strokeStyle to undefined if the path is
filled (not stroked), and set fillStyle to undefined if the path
is stroked only.

Fix assertions in missing_points tests, those were apparently expecting the
inflated numbers.

Remove the strokeStyle='#000000' hack in dygraph-canvas which is now
no longer needed.