Fix tick location according to local time near DST transitions.
The curent algorithm for tick generation uses the 'roll over'
property of the `Date` methods. Hence, it produces two artifacts:

  1.  A one-hour gap of ticks within an hour of the 'fall back' transition.

  2.  Moves ticks within the hour skipped at the 'spring forward' transition.

To solve the first defect, use the constant spacing in milliseconds
for granularities not coarser than hourly.

To solve the second defect, check that computed dates using the 'roll over'
method return the expected tick position.