Dygraph.findPos may return NaN
I embedded a Dygraph into a webcomponent showing several graphs within a elements generated by <template repeat>. Clicks within the graphs did not work correctly. The reason seems to be that the template created a document-fragment around the dygraph element which returned NaN on obj.scrollLeft and obj.scrollTop. The while loop going through all parent nodes up to document.body would pass this document-frament and carry on the NaN returned there.

The consequence was that the x and y returned from Dygraph.findPos resulted in NaN and subsequent calculations failed.