The event `preventDefault` issue
Because of the error: `Consider marking event handler as 'passive' to make the page more responsive`
```
Passive event listeners are a new feature in the DOM spec that enable developers to opt-in to better scroll performance by eliminating the need for scrolling to block on touch and wheel event listeners. Developers can annotate touch and wheel listeners with {passive: true} to indicate that they will never invoke preventDefault.
```
https://stackoverflow.com/a/61248767/5113030

Actually, I don't know if this would be correct, though.
Please, consider looking.