also use the (same!) babel call to translate LICENCE.js
The difference remaining now is that the inline source map for
THIS file, which (only one, perhaps because this file contains
no actual code) browserify reads (hopefully) though fails to
strip, stops containing the absolute source path (in the base64
part) as follows:

-{"version":3,"sources":[],"names":[],"mappings":"","file":"/home/node/dygraphs/LICENCE.js","sourcesContent":[]}
+{"version":3,"sources":[],"names":[],"mappings":"","file":"LICENCE.js","sourcesContent":[]}

Might be useful to map that to LICENSE.txt, even…

The difference in the .min source map is this:

-  "file": "dist/dygraph.min.js",
+  "file": "dygraph.min.js",

This… looks like the right thing as well. Huh.