Warn if points are out of range
Some users may improperly define the range of their axes, leading to
points being drawn outside of the chart area. In the worst cases, if
all points are out of range, the chart will not be drawn at all.

Adding a warning can help users identify this issue and correct it.

2 x points are allowed in order to draw the edges of the chart when
zoomed in.