Fix frame progression in animateSelection_
The callback sent to `repeatAndCleanup()` in `animateSelection_` so far ignored the frame number sent by `repeatAndCleanup()`.

The code was assuming that it will be called `steps` time, which is not true if CPU is saturated and there are longer intervals between frames then expected - frames in the middle will be dropped. `repeatAndCleanup()` guarantees that it will call the callback with last step, but without checking the step number it's impossible to leverage that guarantee.

https://github.com/danvk/dygraphs/pull/534 patched the code to ensure that at least `clearSelection` is called at the end, but it seems like unnecessary complication, and doesn't fix the fade progression if the frames are delayed or irregular.

With current `fadeLevel` computed from the actual step, the cleanup function is no longer necessary.